requirements.txt
flask
tensorflow
pillow
numpy
vercel-wsgi